w0002/requirement.txt
Django==1.11
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.10
mysqlclient==1.3.12
pytz==2017.3
